requirements.txt
fastapi==0.111.0
json5==0.9.25
numpy==1.26.4
pillow==10.3.0
pydantic==2.7.1
pydantic-extra-types==2.7.0
pydantic-settings==2.2.1
pydantic_core==2.18.2
starlette==0.37.2
ultralytics==8.1.26
uvicorn==0.29.0
websockets==12.0